@categorical: require ScalarInt field annotations + ScalarInt runtime values (#349)
The decorator used to read `__annotations__` only for field names and
never validate them — every consumer wrote `field: int` even though
codes flowed through JAX as `jnp.int32` everywhere downstream. The
mismatch was a consistent lie ty couldn't catch, and downstream
`jnp.int32(...)` wraps papered over the dtype gap at runtime.

This change closes the gap at all three layers in lockstep:

* **Annotation gate.** `@categorical` requires every field to be
  annotated `ScalarInt` (from `lcm.typing`); other annotations raise
  the new `CategoricalDefinitionError` at decoration time, naming the
  offending fields and pointing at the import.
* **Runtime values.** Class- and instance-level attribute access
  return 0-d `jnp.int32` scalars. The decorator assigns
  `field(default=i, init=False)` (Python int placeholders for
  `dataclass(frozen=True)`'s mutable-default check) then overrides
  the class attributes with `jnp.int32(i)` via `type.__setattr__`
  post-decoration; `init=False` keeps instance `__dict__` empty so
  attribute lookup falls through to the class scalar.
* **Validator.** `validate_category_class` now checks
  `isinstance(value, jax.Array) and value.shape == () and
  jnp.issubdtype(value.dtype, jnp.integer)`; the consecutiveness
  check coerces via `int(v)`.

Hashability fix: JAX 0-d arrays aren't hashable, so the two pylcm
dict-inversion sites (`simulation/simulate.py`, `initial_conditions.py`)
and `DiscreteGrid.codes` now coerce to Python `int` at the boundary
where hashable keys are needed. New helper
`lcm.invert_regime_ids` lives in `lcm.utils.containers` and centralises
the inversion pattern; downstream consumers can call it instead of
hand-rolling `{int(v): k for k, v in ...}`. `RegimeNamesToIds`
tightens to `MappingProxyType[RegimeName, ScalarInt]`.

Sweep: every `@categorical`-decorated class in `src/lcm_examples`,
`tests/`, `tests/test_models/`, and the user-guide notebooks/docs
moves from `field: int` to `field: ScalarInt`. The `Effort` example
in `mahler_yum_2024/_model.py` switches to the `make_dataclass +
type.__setattr__` shim so its 40 `ScalarInt` defaults bypass
`dataclass`'s mutable-default check. `tests/conftest.py` and
`tests/test_grids.py` gain a `_make_dc` helper for hand-crafted
test dataclasses with `ScalarInt` values.

+++ tests/simulation/test_simulate.py
@@ -39,7 +39,7 @@ def simulate_inputs():
     )
     regimes = {"working_life": updated_working_life, "dead": dead}
     regime_names_to_ids = MappingProxyType(
-        {name: idx for idx, name in enumerate(regimes.keys())}
+        {name: jnp.int32(idx) for idx, name in enumerate(regimes.keys())}
     )
     internal_regimes = process_regimes(
         regimes=regimes,
